# docs/openapi/auth.yaml
openapi: 3.0.3
info:
  title: ID Gateway Auth API
  version: 0.1.0
  description: |
    OpenAPI specification for the ID Gateway authentication endpoints.

    The API follows an OAuth2-style authorization code flow (OIDC-lite) with
    endpoints for initiating authorization, exchanging authorization codes for
    tokens, and fetching user information for an authenticated session.

    **Per-Tenant Issuer (RFC 8414 Compliance):**

    Each tenant has a unique issuer URL following the format: `{base_url}/tenants/{tenant_id}`

    Example: `https://auth.credo.io/tenants/550e8400-e29b-41d4-a716-446655440000`

    The `iss` claim in access tokens and ID tokens uses this per-tenant issuer format.
    The `tenant_id` is also included as a custom claim in access tokens for client convenience.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /auth/authorize:
    post:
      security: []
      summary: Initiate authorization by email
      description: |
        Starts an authorization session for a user. If the email does not exist
        the user record is created automatically. Returns an authorization code
        and the redirect URI that should be used by the caller.

        **Rate Limiting:**
        This endpoint is rate-limited by email + IP to prevent brute force attacks.
        When rate limited, the response includes a `Retry-After` header indicating
        when to retry. After repeated failures, the account may be temporarily locked.

        **Error Handling (RFC 6749 §4.1.2.1):**
        Per RFC 6749, if client_id is unknown/invalid or redirect_uri is mismatched,
        the server MUST NOT redirect and should return an error directly:
        - Unknown client_id → 400 `invalid_client`
        - Inactive client → 400 `invalid_client`
        - redirect_uri not in registered URIs → 400 `bad_request`
        - Inactive tenant → 400 `access_denied`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorizationRequest"
            examples:
              default:
                summary: Minimal authorize request
                value:
                  email: user@example.com
                  client_id: demo-client
                  scopes: ["openid", "profile"]
                  redirect_uri: https://app.example.com/callback
                  state: abc123
      responses:
        "200":
          description: Authorization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationResponse"
              examples:
                default:
                  value:
                    code: authz_cafedeadbeef
                    redirect_uri: https://app.example.com/callback
        "400":
          description: |
            Authorization error (RFC 6749 §4.1.2.1). Check the `error` field:
            - `invalid_client`: Unknown or inactive client_id
            - `access_denied`: Tenant is inactive
            - `bad_request`: Invalid email, redirect_uri format, or redirect_uri not registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_client:
                  value:
                    error: invalid_client
                    error_description: Unknown client_id
                access_denied:
                  value:
                    error: access_denied
                    error_description: Tenant is inactive
                bad_request:
                  value:
                    error: bad_request
                    error_description: redirect_uri not in client's registered URIs
        "429":
          description: |
            Rate limit exceeded. Check the `Retry-After` header for when to retry.
          headers:
            Retry-After:
              description: Seconds until the rate limit resets
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitErrorResponse"
              examples:
                rate_limited:
                  value:
                    error: rate_limit_exceeded
                    message: Too many requests. Please try again later.
                    retry_after: 60
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/token:
    post:
      security: []
      summary: Exchange code or refresh token for new tokens
      description: |
        Exchanges a valid authorization code (FR-1) or refresh token (FR-2)
        for new tokens. For authorization code exchange, the `redirect_uri`
        and `client_id` must match the authorize request. For refresh grant,
        provide a valid refresh token and the `client_id`.

        **Rate Limiting:**
        This endpoint is rate-limited by client_id + IP to prevent credential stuffing.
        When rate limited, the response includes a `Retry-After` header indicating
        when to retry.

        **Error Handling (RFC 6749 §5.2):**
        - Invalid/expired/used authorization code → 400 `invalid_grant`
        - Invalid/expired/used refresh token → 400 `invalid_grant`
        - redirect_uri mismatch → 400 `invalid_grant`
        - Unknown or inactive client → 400 `invalid_client`
        - Missing required fields → 400 `invalid_request`
        - Unsupported grant_type → 400 `unsupported_grant_type`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/TokenRequestAuthCode"
                - $ref: "#/components/schemas/TokenRequestRefresh"
            examples:
              default:
                value:
                  grant_type: authorization_code
                  code: authz_cafedeadbeef
                  redirect_uri: https://app.example.com/callback
                  client_id: demo-client
              refresh_grant:
                value:
                  grant_type: refresh_token
                  refresh_token: ref_7c9e6679-7425-40de-944b-e07fc1f90ae7
                  client_id: demo-client
      responses:
        "200":
          description: Tokens issued successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              examples:
                default:
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    id_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: ref_7c9e6679-7425-40de-944b-e07fc1f90ae7
                    token_type: Bearer
                    expires_in: 3600
                    scope: openid profile
        "400":
          description: |
            OAuth error (RFC 6749 §5.2). Check the `error` field:
            - `invalid_grant`: Invalid/expired/used code or token, redirect_uri mismatch
            - `invalid_client`: Unknown or inactive client
            - `invalid_request`: Missing required fields
            - `unsupported_grant_type`: Grant type not supported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_grant:
                  value:
                    error: invalid_grant
                    error_description: Authorization code expired
                invalid_client:
                  value:
                    error: invalid_client
                    error_description: Unknown client_id
        "429":
          description: |
            Rate limit exceeded. Check the `Retry-After` header for when to retry.
          headers:
            Retry-After:
              description: Seconds until the rate limit resets
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitErrorResponse"
              examples:
                rate_limited:
                  value:
                    error: rate_limit_exceeded
                    message: Too many requests. Please try again later.
                    retry_after: 60
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/revoke:
    post:
      summary: Revoke access or refresh token (logout)
      security: []
      description: |
        Implements RFC 7009 token revocation (FR-3). Accepts either access
        token (JWT) or refresh token (opaque). Idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeRequest"
            examples:
              refresh:
                value:
                  token: ref_7c9e6679-7425-40de-944b-e07fc1f90ae7
                  token_type_hint: refresh_token
      responses:
        "200":
          description: Revocation succeeded or token already invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/sessions:
    get:
      summary: List active sessions for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/sessions/{session_id}:
    delete:
      summary: Revoke a specific session by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
          description: Target session identifier
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeSessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: |
            Forbidden — session does not belong to the authenticated user.
            Note: Non-existent session IDs also return 403 to prevent session enumeration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/logout-all:
    post:
      summary: Revoke all sessions for current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: except_current
          required: false
          schema:
            type: boolean
            default: true
          description: Keep the current session active when true
      responses:
        "200":
          description: Global logout completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutAllResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/userinfo:
    get:
      summary: Get user information for the authenticated session
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User info for the session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
              examples:
                default:
                  value:
                    sub: 7d3e1c6e-7e1c-4b3c-9d7d-31a322b2fbfb
                    email: user@example.com
                    email_verified: true
                    given_name: Jane
                    family_name: Doe
                    name: Jane Doe
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthorizationRequest:
      type: object
      required: [email, client_id, scopes, redirect_uri]
      properties:
        email:
          type: string
          format: email
          description: User email to authorize or create
          maxLength: 255
        client_id:
          type: string
          description: OAuth client identifier
          minLength: 3
          maxLength: 100
        scopes:
          type: array
          minItems: 1
          maxItems: 20
          description: Scopes requested for the session
          items:
            type: string
            maxLength: 100
            description: Requested scope value
        redirect_uri:
          type: string
          format: uri
          description: Redirect URI to return users after authorization
          maxLength: 2048
        state:
          type: string
          description: Opaque value used to maintain state between request and callback
          maxLength: 500
    AuthorizationResponse:
      type: object
      required: [code, redirect_uri]
      properties:
        code:
          type: string
          description: Authorization code for token exchange
        redirect_uri:
          type: string
          format: uri
          description: Redirect URI to invoke after successful authorization
        device_id:
          type: string
          description: Device identifier for device binding (Phase 1 soft launch)
          example: dev_abc123xyz
    TokenRequestAuthCode:
      type: object
      required: [grant_type, code, redirect_uri, client_id]
      properties:
        grant_type:
          type: string
          enum: [authorization_code]
          description: Must be `authorization_code`
        code:
          type: string
          description: Authorization code received from `/auth/authorize`
        redirect_uri:
          type: string
          format: uri
          description: Redirect URI that matches the authorize call
        client_id:
          type: string
          description: OAuth client identifier used during authorization
    TokenRequestRefresh:
      type: object
      required: [grant_type, refresh_token, client_id]
      properties:
        grant_type:
          type: string
          enum: [refresh_token]
          description: Must be `refresh_token`
        refresh_token:
          type: string
          description: Opaque refresh token issued previously
        client_id:
          type: string
          description: OAuth client identifier
    TokenResponse:
      type: object
      description: |
        Token response containing access token, ID token, and optionally refresh token.

        **Token Claims:**
        Both access tokens and ID tokens include:
        - `iss`: Per-tenant issuer URL in format `{base_url}/tenants/{tenant_id}`
        - `aud`: Audience (client identifier)
        - `exp`, `iat`: Expiration and issued-at timestamps

        Access tokens additionally include:
        - `user_id`, `session_id`, `client_id`, `tenant_id`, `scope`, `jti`
      required: [access_token, id_token, expires_in]
      properties:
        access_token:
          type: string
          description: JWT access token for authenticated requests. Contains per-tenant issuer (`iss`) claim.
        id_token:
          type: string
          description: JWT ID token containing authenticated user claims. Contains per-tenant issuer (`iss`) claim.
        refresh_token:
          type: string
          description: New refresh token when using refresh grant (rotation); also issued on initial exchange
        token_type:
          type: string
          description: Token type indicator; always "Bearer"
          example: Bearer
        expires_in:
          type: integer
          description: Token expiration in seconds
          minimum: 1
          example: 3600
        scope:
          type: string
          description: Space-delimited scopes granted
          example: openid profile
    RevokeRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Token to revoke (access or refresh)
        client_id:
          type: string
          description: OAuth client identifier (optional per RFC 7009, required for public clients)
        token_type_hint:
          type: string
          enum: [access_token, refresh_token]
          description: Optional hint for token type
    RevokeResponse:
      type: object
      required: [revoked]
      properties:
        revoked:
          type: boolean
        message:
          type: string
    SessionsResponse:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionItem"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine readable error code
          enum:
            [
              bad_request,
              unauthorized,
              conflict,
              not_found,
              internal_error,
              invalid_consent,
              missing_consent,
              policy_violation,
              registry_timeout,
              # OAuth 2.0 error codes (RFC 6749 §5.2)
              invalid_grant,
              invalid_client,
              unsupported_grant_type,
              invalid_request,
              access_denied,
            ]
        error_description:
          type: string
          description: Human readable explanation of the error
    RateLimitErrorResponse:
      type: object
      required: [error, message, retry_after]
      properties:
        error:
          type: string
          enum: [rate_limit_exceeded]
          description: Error code indicating rate limiting
        message:
          type: string
          description: Human-readable error message
        retry_after:
          type: integer
          description: Seconds until the rate limit resets (also in Retry-After header)
  responses:
    BadRequest:
      description: Invalid or malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation_failure:
              value:
                error: bad_request
                error_description: Invalid JSON in request body
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing_header:
              value:
                error: unauthorized
                error_description: Missing or invalid Authorization header
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            generic:
              value:
                error: internal_error
    # OAuth 2.0 error responses (RFC 6749 §5.2)
...